Externalise property in correct package
